Move tests to "unit" subdir
