+++ tests/Unit/Compiler/OptimizedRewriteTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace ToyWpRouting\Tests\Compiler;
+namespace ToyWpRouting\Tests\Unit\Compiler;
 
 use PHPUnit\Framework\TestCase;
 use RuntimeException;
