+++ tests/Unit/Compiler/RewriteCollectionCompilerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace ToyWpRouting\Tests\Compiler;
+namespace ToyWpRouting\Tests\Unit\Compiler;
 
 use PHPUnit\Framework\TestCase;
 use Spatie\Snapshots\MatchesSnapshots;
