+++ tests/Unit/DefaultInvocationStrategyTest.php
@@ -2,24 +2,23 @@
 
 declare(strict_types=1);
 
-namespace ToyWpRouting\Tests;
+namespace ToyWpRouting\Tests\Unit;
 
 use PHPUnit\Framework\TestCase;
-use ToyWpRouting\InvokerBackedInvocationStrategy;
+use ToyWpRouting\DefaultInvocationStrategy;
 
-// @todo Test with custom resolver set on Invoker instance?
-class InvokerBackedInvocationStrategyTest extends TestCase
+class DefaultInvocationStrategyTest extends TestCase
 {
     public function testInvoke()
     {
-        $invocationStrategy = new InvokerBackedInvocationStrategy();
+        $invocationStrategy = new DefaultInvocationStrategy();
 
         $this->assertSame('testreturnval', $invocationStrategy->invoke(fn () => 'testreturnval'));
     }
 
     public function testInvokeWithCallableResolver()
     {
-        $invocationStrategy = new InvokerBackedInvocationStrategy();
+        $invocationStrategy = new DefaultInvocationStrategy();
         $invocationStrategy->setCallableResolver(function ($potentialCallable) {
             if ('handler' === $potentialCallable) {
                 return fn () => 'modified';
@@ -33,11 +32,14 @@ class InvokerBackedInvocationStrategyTest extends TestCase
 
     public function testInvokeWithContext()
     {
-        $invocationStrategy = new InvokerBackedInvocationStrategy();
+        $invocationStrategy = new DefaultInvocationStrategy();
 
         $this->assertSame(
             'testreturnval',
-            $invocationStrategy->invoke(fn ($append) => 'test' . $append, ['returnval'])
+            $invocationStrategy->invoke(
+                fn ($params) => 'test' . $params['append'],
+                ['append' => 'returnval']
+            )
         );
     }
 }
