+++ tests/Unit/SupportTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace ToyWpRouting\Tests;
+namespace ToyWpRouting\Tests\Unit;
 
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
